requirements.txt
openai==1.3.8
pandas==2.1.3
streamlit==1.28.1
langchain==0.0.340
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-dotenv==1.0.0